--- src/components/Footer.js
@@ -1,10 +1,0 @@
-import React, { Component } from 'react'
-import './footer.scss'
-
-class Footer extends Component {
-  render() {
-    return <footer />
-  }
-}
-
-export default Footer
