+++ src/components/Footer/footer.scss
@@ -1,0 +1,45 @@
+@import '../colors';
+@import '../breakpoints';
+
+.footer {
+  background: $blue;
+  min-height: 320px;
+  color: $text-color-inverted;
+}
+
+.footer__content {
+  padding: 50px 20px;
+  max-width: 768px;
+  margin: 0 auto;
+}
+
+.footer__columns {
+  display: flex;
+  flex-direction: column;
+
+  @include bp(tablet) {
+    flex-direction: row;
+  }
+}
+
+.footer__column {
+  flex: 1 1 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.footer__column--logos {
+  align-items: center;
+  flex-direction: column-reverse;
+  
+  @include bp(tablet) {
+    align-items: flex-end;
+    flex-direction: column;
+  }
+}
+
+.footer__ebdsa-logo {
+  width: 50px;
+  height: 62px;
+  margin: 40px;
+}
