+++ src/components/_colors.scss
@@ -1,3 +1,4 @@
 $red: #ed2c24;
 $blue: #186bb8;
 $text-color: #333;
+$text-color-inverted: #fff;
