--- src/components/footer.scss
@@ -1,5 +1,0 @@
-@import 'colors';
-
-.footer {
-  background: $blue;
-}
